Bug 1195082 - Move args parser from main to cli module
- Created cli module and its associated test_cli.py unittests
- Move args parser from main to cli module
- Moved args parser (which now concerns only defaults handling) tests from
test_main.py to test_cli.py
