+++ tests/unit/test_main.py
@@ -22,13 +22,11 @@ class TestMainCli(unittest.TestCase):
         self.logger = Mock()
 
     @patch('mozregression.main.check_mozregression_version')
-    @patch('mozregression.main.get_defaults')
     @patch('mozlog.structured.commandline.setup_logging')
     @patch('mozregression.main.set_http_session')
     @patch('mozregression.main.ResumeInfoBisectRunner')
     def do_cli(self, argv, BisectRunner, set_http_session,
-               setup_logging, get_defaults, check_mozregression_version):
-        get_defaults.return_value = dict()
+               setup_logging, check_mozregression_version):
         setup_logging.return_value = self.logger
 
         def create_runner(fetch_config, test_runner, options):
@@ -65,40 +63,6 @@ class TestMainCli(unittest.TestCase):
         output = ''.join(output)
         self.assertEqual(exitcode, 0)
         self.assertIn('usage:', output)
-
-    def test_get_erronous_cfg_defaults(self):
-        handle, filepath = tempfile.mkstemp()
-        self.addCleanup(os.unlink, filepath)
-
-        with os.fdopen(handle, 'w') as conf_file:
-            conf_file.write('aaaaaaaaaaa [Defaults]\n')
-
-        with self.assertRaises(SystemExit):
-            main.get_defaults(filepath)
-
-    def test_get_defaults(self):
-        valid_values = {'http-timeout': '10.2',
-                        'persist': '/home/foo/.mozregression',
-                        'bits': '64'}
-
-        handle, filepath = tempfile.mkstemp()
-        conf_default = main.DEFAULT_CONF_FNAME
-
-        self.addCleanup(os.unlink, filepath)
-        self.addCleanup(setattr, main, "DEFAULT_CONF_FNAME", conf_default)
-
-        main.DEFAULT_CONF_FNAME = filepath
-
-        with os.fdopen(handle, 'w') as conf_file:
-            conf_file.write('[Defaults]\n')
-            for key, value in valid_values.iteritems():
-                conf_file.write("%s=%s\n" % (key, value))
-
-        options = main.parse_args(['--bits=32'])
-
-        self.assertEqual(options.http_timeout, 10.2)
-        self.assertEqual(options.persist, '/home/foo/.mozregression')
-        self.assertEqual(options.bits, '32')
 
     def test_without_args(self):
         self.runner.bisect_nightlies.return_value = 0
